// a2a/schema.ts
export interface JSONRPCMessage {
  id?: number | string | null;
  readonly jsonrpc: "2.0";
}

/**
 * Represents a JSON-RPC request object base structure.
 */
export interface JSONRPCRequest extends JSONRPCMessage {
  method: string;
  params?: { [key: string]: any };
}

/**
 * Represents a JSON-RPC 2.0 Error object.
 */
export interface JSONRPCError {
  code: number;
  message: string;
  data?: any; // Reverted from generic D to any
}

/**
 * Represents a JSON-RPC 2.0 Result object (for successful responses).
 * @template R Type of the result field.
 */
export interface JSONRPCResult<R = any> extends JSONRPCMessage {
  result: R;
  error?: never;
}

/**
 * Represents a JSON-RPC 2.0 Error Response object.
 */
export interface JSONRPCErrorResponse extends JSONRPCMessage {
  result?: never;
  error: JSONRPCError | A2AError; // Uses non-generic JSONRPCError and A2AError
}
// ...
export interface JSONParseError extends JSONRPCError {
  code: -32700;
  message: string;
}

export interface InvalidRequestError extends JSONRPCError {
  code: -32600;
  message: string;
}

export interface MethodNotFoundError extends JSONRPCError {
  code: -32601;
  message: string;
}

export interface InvalidParamsError extends JSONRPCError {
  code: -32602;
  message: string;
}

export interface InternalError extends JSONRPCError {
  code: -32603;
  message: string;
}

export interface TaskNotFoundError extends JSONRPCError {
  code: -32001;
  message: string;
}

export interface TaskNotCancelableError extends JSONRPCError {
  code: -32002;
  message: string;
}

export interface PushNotificationNotSupportedError extends JSONRPCError {
  code: -32003;
  message: string;
}

export interface UnsupportedOperationError extends JSONRPCError {
  code: -32004;
  message: string;
}

export interface ContentTypeNotSupportedError extends JSONRPCError {
  code: -32005;
  message: string;
}

export interface InvalidAgentResponseError extends JSONRPCError {
  code: -32006;
  message: string;
}

/**
 * Union of all A2A specific and standard JSON-RPC errors.
 */
export type A2AError =
  | JSONParseError
  | InvalidRequestError
  | MethodNotFoundError
  | InvalidParamsError
  | InternalError
  | TaskNotFoundError
  | TaskNotCancelableError
  | PushNotificationNotSupportedError
  | UnsupportedOperationError
  | ContentTypeNotSupportedError
  | InvalidAgentResponseError;
// ...
export interface SendMessageSuccessResponse extends JSONRPCResult<Message | Task> { }
export type SendMessageResponse = SendMessageSuccessResponse | JSONRPCErrorResponse;

// --- SendStreamingMessage ---
export interface SendStreamingMessageSuccessResponse
  extends JSONRPCResult<Message | Task | TaskStatusUpdateEvent | TaskArtifactUpdateEvent> { }
export type SendStreamingMessageResponse =
  | SendStreamingMessageSuccessResponse
  | JSONRPCErrorResponse;

// --- GetTask ---
export interface GetTaskSuccessResponse extends JSONRPCResult<Task> { }
export type GetTaskResponse = GetTaskSuccessResponse | JSONRPCErrorResponse;

// --- CancelTask ---
export interface CancelTaskSuccessResponse extends JSONRPCResult<Task> { }
export type CancelTaskResponse = CancelTaskSuccessResponse | JSONRPCErrorResponse;

// --- SetTaskPushNotificationConfig ---
export interface SetTaskPushNotificationConfigSuccessResponse
  extends JSONRPCResult<TaskPushNotificationConfig> { }
export type SetTaskPushNotificationConfigResponse =
  | SetTaskPushNotificationConfigSuccessResponse
  | JSONRPCErrorResponse;

// --- GetTaskPushNotificationConfig ---
export interface GetTaskPushNotificationConfigSuccessResponse
  extends JSONRPCResult<TaskPushNotificationConfig> { }
export type GetTaskPushNotificationConfigResponse =
  | GetTaskPushNotificationConfigSuccessResponse
  | JSONRPCErrorResponse;


// === Union Types for A2A Requests/Responses ===

export type A2ARequest =
  | SendMessageRequest
  | SendStreamingMessageRequest
  | GetTaskRequest
  | CancelTaskRequest
  | SetTaskPushNotificationConfigRequest
  | GetTaskPushNotificationConfigRequest
  | TaskResubscriptionRequest;

/**
 * Represents any valid JSON-RPC response defined in the A2A protocol.
 */
export type A2AResponse =
  | SendMessageResponse
  | SendStreamingMessageResponse
  | GetTaskResponse
  | CancelTaskResponse
  | SetTaskPushNotificationConfigResponse
  | GetTaskPushNotificationConfigResponse
  | JSONRPCErrorResponse; // Catch-all for other error responses